requirements.txt
streamlit==1.39.0
pandas==2.2.3
matplotlib==3.9.2
numpy==2.1.3
scipy==1.14.1
openpyxl==3.1.5
groq==0.13.0
python-dotenv==1.0.1
duckdb==1.1.3
